fixed import encoding

+++ s-box/iOS/Les Sexperts/Les Sexperts/Question.m
@@ -23,7 +23,7 @@
     for (NSUInteger i = 0; i < count; ++i) {
         // Select a random element between i and end of array to swap with.
         NSInteger nElements = count - i;
-        NSInteger n = arc4random_uniform(nElements) + i;
+        NSInteger n = arc4random_uniform((uint32_t)nElements) + i;
         [results exchangeObjectAtIndex:i withObjectAtIndex:n];
     }
